refactor(imgcore): merge card/character loaders into unified theme loader
Replace the dual card_loader.go + character_loader.go registry split
with a single theme_loader.go. LoadThemes() scans the unified
assets/theme/ tree and dispatches each subdirectory by ren.json
presence: a manifest triggers LoadCharacterTheme, otherwise
LoadFrameTheme. Both paths produce *theme.Theme consumed by the same
compose pipeline. CardTheme is renamed to FrameTheme,
CardThemeToTheme to FrameThemeToTheme, LoadCardTheme to LoadFrameTheme.
The old CardRegistry/CharacterRegistry interfaces and
NewBuiltinCardRegistry/NewBuiltinCharacterRegistry constructors are
removed. composer.NewThemeRegistry now calls asset.LoadThemes() directly
instead of iterating two separate registries.

将 card_loader.go 与 character_loader.go 的双 registry 架构合并为单一的
theme_loader.go。LoadThemes() 扫描统一目录,按 ren.json 是否存在自动分派
到帧加载或立绘加载路径。CardTheme 重命名为 FrameTheme,旧 registry 接口
和构造函数删除。composer.NewThemeRegistry 直接调用 asset.LoadThemes()。
